+++ bus.h
@@ -1,7 +1,7 @@
 extern uint32_t bus_valid;
 extern uint32_t bus_addr;
-extern uint32_t bus_data_in;
-extern uint32_t bus_data_out;
+extern uint8_t  bus_data_in;
+extern uint8_t  bus_data_out;
 extern uint32_t bus_write_enable;
 extern uint32_t bus_mem_select;
 extern uint32_t bus_io_select;
